create database and seed data
