+++ tests/test_api_routes.py
@@ -1,0 +1,12 @@
+import pytest
+
+
+@pytest.mark.asyncio
+async def test_chat_non_json_415(test_client):
+    """test the chat route with a non-json request"""
+    response = test_client.post("/chat")
+
+    assert response.status_code == 422
+    assert response.headers["Content-Type"] == "application/json"
+    assert response.headers["Content-Length"] == "82"
+    assert b'{"detail":[{"type":"missing"' in response.content
