+++ tests/test_frontend_routes.py
@@ -57,14 +57,3 @@ async def test_assets(test_client):
     assert response.status_code == 200
     assert response.headers["Content-Length"] == str(len(assets_file))
     assert assets_file == response.content
-
-
-@pytest.mark.asyncio
-async def test_chat_non_json_415(test_client):
-    """test the chat route with a non-json request"""
-    response = test_client.post("/chat")
-
-    assert response.status_code == 422
-    assert response.headers["Content-Type"] == "application/json"
-    assert response.headers["Content-Length"] == "82"
-    assert b'{"detail":[{"type":"missing"' in response.content
